requirements.txt
Flask
SQLAlchemy
mysql-connector-python
flask-sqlalchemy